MarkingApp/MarkingApp.Functions/Functions/PythonFunctions/requirements.txt
azure-storage-blob
numpy
opencv-python